Crate markedit

Utilities for editing unstructured markdown documents.

Re-exports

pub extern crate pulldown_cmark;

Structs

OneShot
A Matcher which will only ever return true once.

StartOfNextLine
A Matcher which will match the start of the next top-level element after some inner Matcher matches.

Writer
The output buffer used when rewrite() calls Rewriter::rewrite().

Traits

Matcher
A predicate which can be fed a stream of Events and tell you whether they match a desired condition.

Rewriter
Something which can rewrite events.

Functions

between
Gets all Events between (inclusive) two matchers.

insert_after
Inserts some markdown text after whatever is matched by the Matcher.

link_with_url_containing
Matches the start of a link who's URL contains a certain string.

match_indices
Get an iterator over the indices of matching events.

parse
A convenience function for parsing some text into Events without needing to add pulldown_cmark as an explicit dependency.

rewrite
The whole point.

text
Match an Event::Text event with this exact text.

text_containing
Match an Event::Text event which contains the provided string.