Trait markedit::Matcher

pub trait Matcher {
    fn process_next(&mut self, event: &Event) -> bool;

    fn first_match(&mut self, events: &[Event]) -> Option<usize> { ... }

    fn is_in<'src, I, E>(self, events: I) -> bool
    where
        I: IntoIterator<Item = E> + 'src,
        E: Borrow<Event<'src>>,
        Self: Sized,
    { ... }

    fn then_start_of_next_line(self) -> StartOfNextLine<Self>
    where
        Self: Sized,
    { ... }

    fn fuse(self) -> OneShot<Self>
    where
        Self: Sized,
    { ... }
}

A predicate which can be fed a stream of Events and tell you whether they match a desired condition.

Individual Matchers may choose to return true more than once.

Any function which accepts a Event reference and returns a bool can be used as a Matcher.

fn assert_is_matcher(_: impl Matcher) {}

assert_is_matcher(|ev: &Event<'_>| true);

Required methods

fn process_next(&mut self, event: &Event) -> bool

Loading content...
Provided methods

fn first_match(&mut self, events: &[Event]) -> Option<usize>

Find the index of the first Event which is matched by this predicate.

fn is_in<'src, I, E>(self, events: I) -> bool where
I: IntoIterator<Item = E> + 'src,
E: Borrow<Event<'src>>,
Self: Sized,

Checks whether this Matcher would match anything in a stream of Events.

Examples

let src = "# Heading\nsome text";
let matcher = markedit::text("some text");

assert!(matcher.is_in(markedit::parse(src)));

fn then_start_of_next_line(self) -> StartOfNextLine<Self> where
Self: Sized,

Returns a Matcher which will wait until self matches, then return true at the start of the next top-level element.

fn fuse(self) -> OneShot<Self> where
Self: Sized,

Wraps self in a Matcher which will only ever return true once.

Loading content...